requirements.txt
requests
websockets
python-dotenv
sqlalchemy
aiohttp
flask
plotly
pandas